refactor: Separate some functions into CredentialProvider.js in authenticate.jsx
